requirements.txt
appnope==0.1.2
asgiref==3.4.1
asttokens==2.0.5
autopep8==1.5.7
backcall==0.2.0
backports.zoneinfo==0.2.1
black==21.12b0
certifi==2021.5.30
charset-normalizer==2.0.1
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
executing==0.8.2
idna==3.2
ipython==8.0.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pure-eval==0.2.1
pycodestyle==2.7.0
Pygments==2.11.2
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
stack-data==0.1.3
toml==0.10.2
tomli==1.2.3
traitlets==5.1.1
typing-extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
